requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
pymongo==3.10.1
python-dotenv==0.12.0
Werkzeug==1.0.0
